file check

+++ otp-gen.sh
@@ -16,8 +16,10 @@ rowcount=10
 jd=`date +"%j"`
 otpath="./otp-$fname-$jd.txt"
 echo "output path: " $otpath
+if test -f "$otpath"; then
+    rm $otpath
+fi
 
-rm $otpath
 echo "" >> $otpath
 echo "" >> $otpath
